requirements.txt
python-dotenv
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.59.8
fastapi==0.111.1
langserve==0.2.2
sse_starlette==2.1.2
uvicorn